+++ rasa/core/domain.py
@@ -31,7 +31,7 @@ from rasa.core.constants import (
     SLOT_LISTED_ITEMS,
     DEFAULT_INTENTS,
 )
-from rasa.core.events import SlotSet, UserUttered
+from rasa.shared.core.events import SlotSet, UserUttered
 from rasa.shared.core.slots import Slot, UnfeaturizedSlot, CategoricalSlot
 from rasa.utils.endpoints import EndpointConfig
 from rasa.utils.validation import InvalidYamlFileError, validate_yaml_schema
