+++ rasa/core/test.py
@@ -20,7 +20,7 @@ from rasa.nlu.constants import (
     INTENT,
 )
 from rasa.constants import RESULTS_FILE, PERCENTAGE_KEY
-from rasa.core.events import ActionExecuted, UserUttered
+from rasa.shared.core.events import ActionExecuted, UserUttered
 from rasa.core.trackers import DialogueStateTracker
 from rasa.nlu.training_data.formats.readerwriter import TrainingDataWriter
 from rasa.utils.io import DEFAULT_ENCODING
@@ -276,7 +276,7 @@ class WronglyClassifiedUserUtterance(UserUttered):
 
     def inline_comment(self) -> Text:
         """A comment attached to this event. Used during dumping."""
-        from rasa.core.events import md_format_message
+        from rasa.shared.core.events import md_format_message
 
         predicted_message = md_format_message(
             self.text, self.predicted_intent, self.predicted_entities
@@ -284,7 +284,7 @@ class WronglyClassifiedUserUtterance(UserUttered):
         return f"predicted: {self.predicted_intent}: {predicted_message}"
 
     def as_story_string(self, e2e: bool = True) -> Text:
-        from rasa.core.events import md_format_message
+        from rasa.shared.core.events import md_format_message
 
         correct_message = md_format_message(
             self.text, self.intent.get("name"), self.entities
@@ -388,7 +388,7 @@ def _collect_user_uttered_predictions(
 
 
 def _emulate_form_rejection(partial_tracker: DialogueStateTracker) -> None:
-    from rasa.core.events import ActionExecutionRejected
+    from rasa.shared.core.events import ActionExecutionRejected
 
     rejected_action_name: Text = partial_tracker.active_loop["name"]
     partial_tracker.update(ActionExecutionRejected(rejected_action_name))
