+++ rasa/core/tracker_store.py
@@ -32,7 +32,7 @@ from rasa.core.constants import (
 )
 from rasa.core.conversation import Dialogue
 from rasa.core.domain import Domain
-from rasa.core.events import SessionStarted
+from rasa.shared.core.events import SessionStarted
 from rasa.core.trackers import ActionExecuted, DialogueStateTracker, EventVerbosity
 import rasa.cli.utils as rasa_cli_utils
 from rasa.nlu.constants import INTENT_NAME_KEY
