+++ rasa/core/training/interactive.py
@@ -33,8 +33,8 @@ from rasa.core.constants import (
     UTTER_PREFIX,
 )
 from rasa.core.domain import Domain
-import rasa.core.events
-from rasa.core.events import (
+import rasa.shared.core.events
+from rasa.shared.core.events import (
     ActionExecuted,
     ActionReverted,
     BotUttered,
@@ -786,7 +786,7 @@ def _write_stories_to_file(
     with open(export_story_path, append_write, encoding=io_utils.DEFAULT_ENCODING) as f:
         i = 1
         for conversation in sub_conversations:
-            parsed_events = rasa.core.events.deserialise_events(conversation)
+            parsed_events = rasa.shared.core.events.deserialise_events(conversation)
             tracker = DialogueStateTracker.from_events(
                 f"interactive_story_{i}", evts=parsed_events, slots=domain.slots
             )
@@ -1325,7 +1325,7 @@ async def _fetch_events(
             events = tracker.get("events", [])
 
             for conversation in _split_conversation_at_restarts(events):
-                parsed_events = rasa.core.events.deserialise_events(conversation)
+                parsed_events = rasa.shared.core.events.deserialise_events(conversation)
                 event_sequences.append(parsed_events)
         else:
             event_sequences.append(conversation_id)
