+++ rasa/core/training/story_conflict.py
@@ -4,7 +4,7 @@ from typing import List, Optional, Dict, Text, Tuple, Generator, NamedTuple
 
 from rasa.core.actions.action import ACTION_LISTEN_NAME
 from rasa.core.domain import PREV_PREFIX, Domain
-from rasa.core.events import ActionExecuted, Event
+from rasa.shared.core.events import ActionExecuted, Event
 from rasa.core.featurizers import MaxHistoryTrackerFeaturizer
 from rasa.nlu.constants import INTENT
 from rasa.core.training.generator import TrackerWithCachedStates
