+++ rasa/core/training/story_reader/story_step_builder.py
@@ -4,7 +4,7 @@ from typing import Text, Optional, Dict, Any, List
 import rasa.shared.utils.io
 from rasa.constants import DOCS_URL_STORIES
 from rasa.core import utils
-from rasa.core.events import UserUttered
+from rasa.shared.core.events import UserUttered
 from rasa.core.training.structures import (
     Checkpoint,
     GENERATED_CHECKPOINT_PREFIX,
