+++ rasa/server.py
@@ -36,7 +36,7 @@ from rasa.core.channels.channel import (
     UserMessage,
 )
 from rasa.core.domain import InvalidDomain
-from rasa.core.events import Event
+from rasa.shared.core.events import Event
 from rasa.core.lock_store import LockStore
 from rasa.core.test import test
 from rasa.core.tracker_store import TrackerStore
