+++ rasa/validator.py
@@ -14,8 +14,8 @@ from rasa.constants import (
 )
 from rasa.core.constants import UTTER_PREFIX
 from rasa.core.domain import Domain
-from rasa.core.events import ActionExecuted
-from rasa.core.events import UserUttered
+from rasa.shared.core.events import ActionExecuted
+from rasa.shared.core.events import UserUttered
 from rasa.core.training.generator import TrainingDataGenerator
 from rasa.core.training.structures import StoryGraph
 from rasa.importers.importer import TrainingDataImporter
