+++ tests/core/actions/test_loops.py
@@ -4,7 +4,7 @@ import pytest
 from rasa.core.actions.loops import LoopAction
 from rasa.core.channels import CollectingOutputChannel
 from rasa.core.domain import Domain
-from rasa.core.events import (
+from rasa.shared.core.events import (
     Event,
     ActionExecutionRejected,
     ActionExecuted,
