+++ tests/core/actions/test_two_stage_fallback.py
@@ -11,7 +11,7 @@ from rasa.core.actions.two_stage_fallback import (
 from rasa.core.channels import CollectingOutputChannel
 from rasa.core.constants import USER_INTENT_OUT_OF_SCOPE
 from rasa.core.domain import Domain
-from rasa.core.events import (
+from rasa.shared.core.events import (
     ActionExecuted,
     UserUttered,
     ActiveLoop,
