+++ tests/core/test_actions.py
@@ -32,7 +32,7 @@ from rasa.core.actions.forms import FormAction
 from rasa.core.actions.two_stage_fallback import ACTION_TWO_STAGE_FALLBACK_NAME
 from rasa.core.channels import CollectingOutputChannel
 from rasa.core.domain import Domain, SessionConfig
-from rasa.core.events import (
+from rasa.shared.core.events import (
     Restarted,
     SlotSet,
     UserUtteranceReverted,
