+++ tests/core/test_exporter.py
@@ -12,7 +12,7 @@ import rasa.utils.io as io_utils
 from rasa.core.brokers.pika import PikaEventBroker
 from rasa.core.brokers.sql import SQLEventBroker
 from rasa.core.constants import RASA_EXPORT_PROCESS_ID_HEADER_NAME
-from rasa.core.events import SessionStarted, ActionExecuted
+from rasa.shared.core.events import SessionStarted, ActionExecuted
 from rasa.core.tracker_store import SQLTrackerStore
 from rasa.core.trackers import DialogueStateTracker
 from rasa.exceptions import (
