+++ tests/core/test_processor.py
@@ -16,7 +16,7 @@ from rasa.core.actions.action import ACTION_LISTEN_NAME, ACTION_SESSION_START_NA
 from rasa.core.agent import Agent
 from rasa.core.channels.channel import CollectingOutputChannel, UserMessage
 from rasa.core.domain import SessionConfig, Domain
-from rasa.core.events import (
+from rasa.shared.core.events import (
     ActionExecuted,
     BotUttered,
     ReminderCancelled,
