+++ tests/core/test_structures.py
@@ -1,6 +1,6 @@
 from rasa.core.actions.action import ACTION_SESSION_START_NAME
 from rasa.core.domain import Domain
-from rasa.core.events import SessionStarted, SlotSet, UserUttered, ActionExecuted
+from rasa.shared.core.events import SessionStarted, SlotSet, UserUttered, ActionExecuted
 from rasa.core.trackers import DialogueStateTracker
 from rasa.core.training.structures import Story
