+++ tests/core/training/story_reader/test_common_story_reader.py
@@ -9,7 +9,7 @@ import pytest
 
 from rasa.core import training
 from rasa.core.domain import Domain
-from rasa.core.events import UserUttered, ActionExecuted, SessionStarted
+from rasa.shared.core.events import UserUttered, ActionExecuted, SessionStarted
 from rasa.core.featurizers import (
     MaxHistoryTrackerFeaturizer,
     BinarySingleStateFeaturizer,
