+++ tests/core/training/story_reader/test_markdown_story_reader.py
@@ -6,7 +6,7 @@ import pytest
 import rasa.utils.io
 from rasa.core import training
 from rasa.core.domain import Domain
-from rasa.core.events import (
+from rasa.shared.core.events import (
     UserUttered,
     ActionExecuted,
     ActionExecutionRejected,
@@ -269,7 +269,7 @@ async def test_read_stories_with_rules(default_domain: Domain):
 
 async def test_read_rules_without_stories(default_domain: Domain):
     story_steps = await loading.load_data_from_files(
-        ["data/test_stories/rules_without_stories.md"], default_domain,
+        ["data/test_stories/rules_without_stories.md"], default_domain
     )
 
     # this file contains three rules and two ML stories
@@ -305,7 +305,7 @@ async def test_read_rules_without_stories(default_domain: Domain):
     "line, expected",
     [
         (" greet: hi", {"intent": "greet", "text": "hi"}),
-        (" greet: /greet", {"intent": "greet", "text": "/greet", "entities": [],},),
+        (" greet: /greet", {"intent": "greet", "text": "/greet", "entities": []}),
         (
             'greet: /greet{"test": "test"}',
             {
