+++ tests/core/training/story_reader/test_yaml_story_reader.py
@@ -9,7 +9,7 @@ from rasa.core import training
 from rasa.core.actions.action import RULE_SNIPPET_ACTION_NAME
 from rasa.core.domain import Domain
 from rasa.core.training import loading
-from rasa.core.events import ActionExecuted, UserUttered, SlotSet, ActiveLoop
+from rasa.shared.core.events import ActionExecuted, UserUttered, SlotSet, ActiveLoop
 from rasa.core.training.story_reader.yaml_story_reader import YAMLStoryReader
 from rasa.core.training.structures import StoryStep
