+++ tests/core/utilities.py
@@ -2,14 +2,14 @@ import itertools
 
 import contextlib
 import typing
-from typing import Text, List, Optional, Text, Any, Dict
+from typing import List, Optional, Text, Any, Dict
 
 import jsonpickle
 import os
 
 import rasa.utils.io
 from rasa.core.domain import Domain
-from rasa.core.events import UserUttered, Event
+from rasa.shared.core.events import UserUttered, Event
 from rasa.core.trackers import DialogueStateTracker
 from rasa.nlu.constants import INTENT_NAME_KEY
 from tests.core.conftest import DEFAULT_DOMAIN_PATH_WITH_SLOTS
