+++ tests/test_server.py
@@ -5,7 +5,6 @@ from unittest.mock import Mock, ANY
 
 import requests
 import time
-import tempfile
 import uuid
 
 from typing import List, Text, Type, Generator, NoReturn, Dict
@@ -23,11 +22,12 @@ import rasa
 import rasa.constants
 import rasa.utils.io
 import rasa.server
-from rasa.core import events, utils
+from rasa.core import utils
+from rasa.shared.core import events
 from rasa.core.agent import Agent
 from rasa.core.channels import CollectingOutputChannel, RestInput, SlackInput
 from rasa.core.channels.slack import SlackBot
-from rasa.core.events import Event, UserUttered, SlotSet, BotUttered
+from rasa.shared.core.events import Event, UserUttered, SlotSet, BotUttered
 from rasa.core.trackers import DialogueStateTracker
 from rasa.model import unpack_model
 from rasa.nlu.constants import INTENT_NAME_KEY
